requirements.txt
async-timeout==4.0.3
bleach==6.1.0
blinker==1.7.0
cachelib==0.13.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.2
flask-cors==5.0.1
Flask-Session==0.8.0
Flask-SSLify==0.1.5
flask-talisman==1.1.0
gunicorn==21.2.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
msgspec==0.19.0
oauthlib==3.2.2
packaging==24.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-discogs-client==2.7
RapidFuzz==3.13.0
redis==5.0.3
requests==2.31.0
six==1.16.0
spotipy==2.23.0
urllib3==2.2.1
webencodings==0.5.1
Werkzeug==3.0.1
